+++ src/tests/channel.js
@@ -1,0 +1,11 @@
+import basicFunctions from './channel/basicFunctions';
+import adapter from './channel/adapter';
+
+export default function () {
+    describe('Class Channel:', () => {
+        describe('Function check:', () => {
+            adapter();
+            basicFunctions();
+        });
+    });
+}