+++ src/tests/index.js
@@ -1,0 +1,7 @@
+require('source-map-support').install();
+
+import ChannelTest from './channel';
+
+describe('AncientSouls/Channels:', () => {
+    ChannelTest();
+});